refactor: fix typo in name of useMostRecentVehicleById.ts
