+++ assets/src/hooks/useMostRecentVehicleById.ts
